requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==41.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.103.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.13.0
uvicorn==0.23.2
